Optimize DateRange setter by keeping an internal cache.
This allows the speed of the matrix calculator to be more than double that it was before.

An example benchmark: over a span of 20 years, 40k randomized modactions, (via some DB manipulation) and the day rangetype, calculation takes 29.9 seconds Given the fact that in reality there would be more dates inserted into the cache because the TTL of modactions is only 120/90 days, it should theoretically take less than 10 seconds. The week calulator, as it is, only takes ~7 seconds.

This also keeps the datekeys of the matrix in a unform order.

The custom key type can not be cached in this method, and may be worth invetigating a different caching method for it.

cc @noeatnosleep