Mention karma in CANT_CREATE_SR
The most common, and I really mean it, the most common post in /r/help and related subreddits is when users question why they can't make a subreddit even though they have an old account. The error that users get only mentions account age and does not mention any kind of karma requirement. It's in the /r/help FAQ, but we all know no one reads that.

Just lightly mentioning karma could reduce the traffic on /r/help as well as be more clear.